+++ sdk/resourcemanager/fabric/armfabric/responses.go
@@ -51,6 +51,12 @@ type CapacitiesClientListSKUsResponse struct {
 	RpSKUEnumerationForNewResourceResult
 }
 
+// CapacitiesClientListUsagesResponse contains the response from method CapacitiesClient.NewListUsagesPager.
+type CapacitiesClientListUsagesResponse struct {
+	// An object that represents enumerating usage and quota details for existing resources
+	RpUsageAndQuotaEnumerationForExistingResourceResult
+}
+
 // CapacitiesClientResumeResponse contains the response from method CapacitiesClient.BeginResume.
 type CapacitiesClientResumeResponse struct {
 	// placeholder for future response values